completed the eighth task "gc" fix 1.2
