+++ module1/src/main/java/net/broscorp/gc/FirstTest.java
@@ -1,0 +1,48 @@
+package net.broscorp.gc;
+
+import org.testng.annotations.Test;
+
+public class FirstTest {
+
+  public class FirstClass {
+
+    private SecondClass secondClass;
+
+    @Override
+    protected void finalize() throws Throwable {
+      System.out.println("The memory was freed. Object A has been deleted");
+      super.finalize();
+    }
+
+    public void secondClassSet(SecondClass secondClass) {
+      this.secondClass = secondClass;
+    }
+  }
+
+  public class SecondClass {
+
+    private FirstClass firstClass;
+
+    @Override
+    protected void finalize() throws Throwable {
+      System.out.println("The memory was freed. Object B has been deleted");
+      super.finalize();
+    }
+
+    public void setFirstClassfirstClassSet(FirstClass firstClass) {
+      this.firstClass = firstClass;
+    }
+  }
+
+  @Test
+  public void createFirstAndSecondClass() {
+    for (int i = 0; i < 100_000; i++) {
+      FirstClass firstClass = new FirstClass();
+      SecondClass secondClass = new SecondClass();
+      firstClass.secondClass = secondClass;
+      secondClass.firstClass = firstClass;
+    }
+    System.gc();
+  }
+
+}
